+++ .changeset/short-zoos-sniff.md
@@ -1,0 +1,5 @@
+---
+"@thirdweb-dev/service-utils": patch
+---
+
+exported usageEventSchema for cf worker and also UsageEvent for node
