+++ packages/service-utils/src/node/index.ts
@@ -11,6 +11,7 @@ import type { AuthorizationInput } from "../core/authorize";
 import type { AuthorizationResult } from "../core/authorize/types";
 import type { CoreAuthInput } from "../core/types";
 
+export * from "../core/usage";
 export * from "../core/rateLimit";
 export * from "../core/services";
 export * from "../core/usageLimit";
